given bigger data type

+++ codechef/XORAGN.java
@@ -5,14 +5,14 @@ class XORAGN{
 
     public static void main(String[] args){
         Scanner sc = new Scanner(System.in);
-        int t = sc.nextInt();
+        long t = sc.nextInt();
         for(int testCase=0; testCase < t ; testCase++){
             int size = sc.nextInt();
-            int[] a = new int[size];
+            long[] a = new long[size];
             // int[] b = new int[size*size];
-            int x;
+            long x;
             // int k = 0;
-            int ans=0;
+            long ans=0;
             for(int i=0; i<size; i++){
                 a[i] = sc.nextInt();
             }
